+++ R/autoplot-density.R
@@ -87,7 +87,7 @@ tidy_autoplot <- function(.data, .plot_type = "density", .line_size = .5,
     "tidy_zero_truncated_poisson", "tidy_zero_truncated_geometric",
     "tidy_zero_truncated_binomial", "tidy_zero_truncated_negative_binomial",
     "tidy_pareto_single_parameter", "tidy_pareto", "tidy_inverse_pareto",
-    "tidy_generalized_pareto"
+    "tidy_generalized_pareto","tidy_paralogistic"
   )) {
     rlang::abort("The data passed must come from a `tidy_` distribution function.")
   }
@@ -160,6 +160,10 @@ tidy_autoplot <- function(.data, .plot_type = "density", .line_size = .5,
         "Rate: ", atb$.rate, " - ",
         "Scale: ", atb$.scale
       )
+    } else if (atb$tibble_type %in% c("tidy_paralogistic")){
+      paste0("Shape: ", atb$.shape, " - ",
+             "Rate: ", atb$.rate, " - ",
+             "Scale: ", atb$.scale)
     }
   )